requirements.txt
Flask==2.2.5
Werkzeug==2.3.0
flask-smorest
python-dotenv
sqlalchemy
flask-sqlalchemy
flask-jwt-extended
passlib
flask-migrate
gunicorn
psycopg2
requests
rq
redis